+++ lib/ClangImporter/CMakeLists.txt
@@ -13,6 +13,7 @@ add_swift_host_library(swiftClangImporter STATIC
   ClangImporterRequests.cpp
   ClangModuleDependencyScanner.cpp
   ClangSourceBufferImporter.cpp
+  SwiftDeclSynthesizer.cpp
   DWARFImporter.cpp
   ImportDecl.cpp
   ImportEnumInfo.cpp
